% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubeau_piezo.R
\name{chronique_piezo}
\alias{chronique_piezo}
\title{Téléchargement des chronique piézomètriques}
\usage{
chronique_piezo(id = "07548X0009/F")
}
\arguments{
\item{id}{ancien code BSS}
}
\value{
dataframe
}
\description{
Téléchargement des chronique piézomètriques
}
